+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### 🧹 Housekeeping
+
+- Replaced Moq with Nsubstitute ([#117](https://github.com/candoumbe/DataFilters.AspNetCore/issues/117))
 
 ## [0.4.0] / 2022-08-17
 - Moved `MinimalAttribute` to a dedicated [`DataFilters.AspNetCore.Attributes`](https://nuget.org/packages/DataFilters.AspNetCore.Attributes) nuget package